Splitted state machine; Linker problems
