+++ vs/driver64/driver64.vcxproj
@@ -161,6 +161,7 @@
     <ClInclude Include="..\..\driver\environment.h" />
     <ClInclude Include="..\..\driver\escaping\escape_sequences.h" />
     <ClInclude Include="..\..\driver\escaping\lexer.h" />
+    <ClInclude Include="..\..\driver\escaping\state_machine.h" />
     <ClInclude Include="..\..\driver\escaping\string_view.h" />
     <ClInclude Include="..\..\driver\log.h" />
     <ClInclude Include="..\..\driver\platform.h" />
@@ -180,6 +181,7 @@
     <ClCompile Include="..\..\driver\environment.cpp" />
     <ClCompile Include="..\..\driver\escaping\escape_sequences.cpp" />
     <ClCompile Include="..\..\driver\escaping\lexer.cpp" />
+    <ClCompile Include="..\..\driver\escaping\state_machine.cpp" />
     <ClCompile Include="..\..\driver\handles.cpp" />
     <ClCompile Include="..\..\driver\info.cpp" />
     <ClCompile Include="..\..\driver\odbc.cpp" />
